--- bloptools/gp/configs.py
@@ -1,13 +1,0 @@
-CONFIGS = {
-    "tes" : {
-        "fields" : ['vstream_image', 'I0'],
-        "fitness_args" : {
-            'image' : 'vstream_image',
-             'flux' : 'I0',
-   'target_width_x' : 0, 
-   'target_width_y' : 0,
-        'beam_prop' : 0.5,
- 'min_separability' : 0.5,
-   'OOB_rel_buffer' : 1/32, }
-    },
-}
